# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Swagger with Laravel",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/parts": {
            "post": {
                "tags": [
                    "Part"
                ],
                "summary": "Store a newly created part",
                "operationId": "f1730ccc6cc559a32452c97a747d4bbf",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "part_number": {
                                        "type": "string",
                                        "example": "PN123"
                                    },
                                    "task_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "appliance_id": {
                                        "type": "integer",
                                        "example": 2
                                    },
                                    "distributor_id": {
                                        "type": "integer",
                                        "example": 3
                                    },
                                    "status_id": {
                                        "type": "integer",
                                        "example": 4
                                    },
                                    "qnt": {
                                        "type": "integer",
                                        "example": 5
                                    },
                                    "dealer_price": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 100
                                    },
                                    "retail_price": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 150
                                    },
                                    "distributor_name": {
                                        "type": "string",
                                        "example": "Distributor Inc."
                                    },
                                    "part_description": {
                                        "type": "string",
                                        "example": "A high-quality part for repair."
                                    },
                                    "distributor_invoice": {
                                        "type": "string",
                                        "example": "INV-123456"
                                    },
                                    "eta": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2024-08-01"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Part created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/PartResource"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Failed to store part",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Failed to store part."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/parts/{part}": {
            "put": {
                "tags": [
                    "Part"
                ],
                "summary": "Update the specified part",
                "operationId": "06c954a371adcc3c9022a6ab725b2a41",
                "parameters": [
                    {
                        "name": "part",
                        "in": "path",
                        "description": "ID of the part to update",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "part_number": {
                                        "type": "string",
                                        "example": "PN123"
                                    },
                                    "task_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "appliance_id": {
                                        "type": "integer",
                                        "example": 2
                                    },
                                    "distributor_id": {
                                        "type": "integer",
                                        "example": 3
                                    },
                                    "status_id": {
                                        "type": "integer",
                                        "example": 4
                                    },
                                    "qnt": {
                                        "type": "integer",
                                        "example": 5
                                    },
                                    "dealer_price": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 100
                                    },
                                    "retail_price": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 150
                                    },
                                    "distributor_name": {
                                        "type": "string",
                                        "example": "Distributor Inc."
                                    },
                                    "part_description": {
                                        "type": "string",
                                        "example": "A high-quality part for repair."
                                    },
                                    "distributor_invoice": {
                                        "type": "string",
                                        "example": "INV-123456"
                                    },
                                    "eta": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2024-08-01"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Part updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/PartResource"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Failed to update part",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Failed to update part."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Part"
                ],
                "summary": "Remove the specified part",
                "operationId": "f60df46d6107d6b726a398d84c90f84f",
                "parameters": [
                    {
                        "name": "part",
                        "in": "path",
                        "description": "ID of the part to delete",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Part deleted successfully"
                    },
                    "500": {
                        "description": "Failed to delete part",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Failed to delete part."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "PartResource": {
                "required": [
                    "id",
                    "part_number_id",
                    "task_id",
                    "appliance_id",
                    "distributor_id",
                    "user_id",
                    "status_id",
                    "qnt",
                    "dealer_price",
                    "retail_price"
                ],
                "properties": {
                    "id": {
                        "description": "The unique identifier for the part.",
                        "type": "integer",
                        "example": 1
                    },
                    "part_number_id": {
                        "description": "The ID of the part number.",
                        "type": "integer",
                        "example": 101
                    },
                    "task_id": {
                        "description": "The ID of the associated task.",
                        "type": "integer",
                        "example": 202
                    },
                    "appliance_id": {
                        "description": "The ID of the associated appliance.",
                        "type": "integer",
                        "example": 303
                    },
                    "distributor_id": {
                        "description": "The ID of the distributor.",
                        "type": "integer",
                        "example": 404
                    },
                    "user_id": {
                        "description": "The ID of the user associated with the part.",
                        "type": "integer",
                        "example": 505
                    },
                    "status_id": {
                        "description": "The ID of the part status.",
                        "type": "integer",
                        "example": 606
                    },
                    "qnt": {
                        "description": "The quantity of parts.",
                        "type": "integer",
                        "example": 10
                    },
                    "dealer_price": {
                        "description": "The dealer price of the part.",
                        "type": "number",
                        "format": "float",
                        "example": 100
                    },
                    "retail_price": {
                        "description": "The retail price of the part.",
                        "type": "number",
                        "format": "float",
                        "example": 150
                    },
                    "distributor_name": {
                        "description": "The name of the distributor.",
                        "type": "string",
                        "maxLength": 255,
                        "example": "Distributor Inc."
                    },
                    "part_description": {
                        "description": "A description of the part.",
                        "type": "string",
                        "example": "A high-quality part for repair."
                    },
                    "distributor_invoice": {
                        "description": "The distributor's invoice number.",
                        "type": "string",
                        "example": "INV-123456"
                    },
                    "eta": {
                        "description": "The estimated time of arrival of the part.",
                        "type": "string",
                        "format": "date",
                        "example": "2024-08-01"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}